apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

argon2-cffi==23.1.0
PyJWT==2.10.1

google-cloud-storage==2.18.2
google-cloud-pubsub==2.27.1
google-auth==2.36.0

stripe==11.4.1
requests==2.32.3

python-multipart==0.0.20




